Extract common EIS settings utils
