+++ x-pack/plugin/inference/src/main/java/org/elasticsearch/xpack/inference/services/ServiceFields.java
@@ -34,7 +34,6 @@ public final class ServiceFields {
      * {@link org.elasticsearch.inference.TaskType#EMBEDDING} task type. Defaults to true.
      */
     public static final String MULTIMODAL_MODEL = "multimodal_model";
-    public static final String MAX_BATCH_SIZE = "max_batch_size";
 
     private ServiceFields() {
